+++ src/types/types.ts
@@ -1239,6 +1239,21 @@ export interface EdgeCurrencyWallet {
   readonly otherMethods: EdgeOtherMethods
 }
 
+export interface EdgeMemoryWallet {
+  readonly watch: Subscriber<EdgeMemoryWallet>
+  readonly balanceMap: EdgeBalanceMap
+  readonly detectedTokenIds: string[]
+  readonly syncRatio: number
+  readonly changeEnabledTokenIds: (tokenIds: string[]) => Promise<void>
+  readonly startEngine: () => Promise<void>
+  readonly getMaxSpendable: (spendInfo: EdgeSpendInfo) => Promise<string>
+  readonly makeSpend: (spendInfo: EdgeSpendInfo) => Promise<EdgeTransaction>
+  readonly signTx: (tx: EdgeTransaction) => Promise<EdgeTransaction>
+  readonly broadcastTx: (tx: EdgeTransaction) => Promise<EdgeTransaction>
+  readonly saveTx: (tx: EdgeTransaction) => Promise<void>
+  readonly close: () => Promise<void>
+}
+
 // ---------------------------------------------------------------------
 // swap plugin
 // ---------------------------------------------------------------------
@@ -1636,6 +1651,10 @@ export interface EdgeAccount {
     walletId: string
   ) => Promise<EdgeCurrencyWallet>
   readonly waitForAllWallets: () => Promise<void>
+  readonly makeMemoryWallet: (
+    walletType: string,
+    opts?: EdgeCreateCurrencyWalletOptions
+  ) => Promise<EdgeMemoryWallet>
 
   // Token & wallet activation:
   readonly getActivationAssets: (